requirements.txt
streamlit
langchain
chromadb
pandas
matplotlib
plotly
google-generativeai
pypdf
python-dotenv
